Use Enum Maps instead of HashMaps

+++ builder-core/src/bonus/bonus_source.rs
@@ -1,11 +1,20 @@
+use enum_map::Enum;
+
 use crate::{attribute::Attribute, feat::Feat};
 
-#[derive(Clone, Copy, PartialEq, Eq, Hash, serde::Serialize, serde::Deserialize, Debug)]
+/// Describes the source that a bonus is from.
+///
+/// Sources can be either from an [Attribute](crate::bonus::BonusSource::Attribute(crate::attribute::Attribute)), [Feat](crate::bonus::BonusSource::Feat(crate::feat::Feat)), many others to come, and lastly [Unique](crate::bonus::BonusSource::Unique(u8))
+#[derive(Clone, Copy, PartialEq, Eq, Hash, serde::Serialize, serde::Deserialize, Debug, Enum)]
 pub enum BonusSource {
+    /// Represents any source from an [Attribute].
+    ///
+    /// This is the most common source, as it envelops Feats and more through attributes like [Attribute::Feat]
     Attribute(Attribute),
+    /// Represents sources from a [Feat]
     Feat(Feat),
-    /// Indicates some unique identifier, indicated with a usize
-    Unique(usize),
+    /// Indicates some unique identifier, indicated with a [u8]
+    Unique(u8),
 }
 
 impl ToString for BonusSource {
